score; enemies; win conditions; Missile Direction
Implemented a lot with this push. Enemies were changed so that one enemy
object could change its sprite and depending on the sprite would have
different characteristics (speed, health etc). Score now counts up at
the end of each level. When you complete the win conditions for the
level the game is over and it goes to the score screen. Or if you lose
all your health the game is over and you cannot continue to the next
level. One final thing is I used the mouse and player's position to
calculate a unit vector then multiplied it by the speed of the missile
to go in that particular direction.