Fixed Unit Vectors on Mouse Position
I had previously just used the two vectors that were produced and found
unit vectors based on the two (x and y) adding up to 1. This logic,
while it worked, slowed the missile down when it wasn't going straight
in the x and y direction so I adjusted it to use the magnitude of the
original line to find the unit vectors.